Ask for already installed ZeroNet path at first launch, added config file, detect it ZeroNet is already running.

+++ launch.py
@@ -1,21 +1,91 @@
 #!/usr/bin/python
 
 import sys
-import os.path
+import os
+import ConfigParser
 from MainWindow import MainWindow
-from ZeroNet import zeronet
 from PyQt5.QtWidgets import QApplication
 from multiprocessing import Process, freeze_support
+from ZeroNet import zeronet
 import time
+import imp
+
+# See if it is lock or not
+def openLocked(path, mode="w"):
+    if os.name == "posix":
+        import fcntl
+        f = open(path, mode)
+        fcntl.flock(f, fcntl.LOCK_EX | fcntl.LOCK_NB)
+    elif os.name == "nt":
+        import msvcrt
+        f = open(path, mode)
+        msvcrt.locking(f.fileno(), msvcrt.LK_NBLCK, -1)
+    else:
+        f = open(path, mode)
+    return f
 
 if __name__ == '__main__':
     freeze_support()
-    print "Inside main !"
-    # Create a process for Zeronet
-    p = Process(target=zeronet.main)
-    print "Created Process"
-    p.start()
-    print "Starting Process"
+
+    p = None
+
+    config = ConfigParser.ConfigParser()
+    config.read('browser.conf')
+
+    if not config.has_section('global'):
+        config.add_section('global')
+
+        answer = raw_input("Do you already have ZeroNet installed somewhere ? (Y\\N) \n")
+
+        if (answer == 'Y'):
+            home = os.path.expanduser("~")
+            zeronet_path = False
+            for root, dirs, files in os.walk(home):
+                if not root.startswith('.'):
+                    for dir in dirs:
+                        if dir.startswith('ZeroNet'):
+                            path = os.path.join(root, dir)
+                            answer = raw_input("Is it this path correct "+path+" ?")
+                            if (answer == 'Y'):
+                                if (os.path.exists(path) and os.path.isdir(path)):
+                                    zeronet_path = path
+                                    break
+                if zeronet_path:
+                    config.set('global', 'zeronet_path', zeronet_path)
+                    # Create a __init__.py file
+                    open(os.path.join(zeronet_path, '__init__.py'), 'w')
+                    break
+        else:
+            config.set('global', 'zeronet_path', '')
+
+        with open('browser.conf', 'wb') as configfile:
+            config.write(configfile)
+
+        print "Please restart to load the config"
+        sys.exit(0)
+    else:
+        zeronet_path = config.get('global', 'zeronet_path')
+        if zeronet_path:
+            try:
+                zeronet = imp.load_source('zeronet', os.path.join(zeronet_path, 'zeronet.py'))
+            except:
+                print "Error - Couldn't load ZeroNet from given path. Loading local ZeroNet."
+
+    if zeronet_path:
+        # See if it is already running
+        try:
+            print "OK"
+            lock = openLocked("%s/lock.pid" % os.path.join(zeronet_path,'data'), "w")
+            lock.close()
+            # Create a process for Zeronet using this version of ZeroNet
+            p = Process(target=zeronet.main)
+            p.start()
+        except IOError as err:
+            print "Can't open lock file, your ZeroNet client is probably already running, opening browser without starting Zeronet in the background..."
+    else:
+        # Create a process for Zeronet
+        p = Process(target=zeronet.main)
+        p.start()
 
     time.sleep(5)
 
@@ -24,5 +94,6 @@ if __name__ == '__main__':
     mainWindow = MainWindow()
     app.exec_()
 
-    # Shutdown Zeronet
-    p.terminate()
+    if p:
+        # Shutdown Zeronet if runingin the background
+        p.terminate()
